output_path -> image_path + Quantitiy refactor
