+++ autolens/analysis/plotter_interface.py
@@ -22,7 +22,7 @@ class PlotterInterface(AgPlotterInterface):
 
     Parameters
     ----------
-    output_path
+    image_path
         The path on the hard-disk to the `image` folder of the non-linear searches results.
     """
 
@@ -30,8 +30,8 @@ class PlotterInterface(AgPlotterInterface):
         """
         Visualizes a `Tracer` object.
 
-        Images are output to the `image` folder of the `output_path` in a subfolder called `tracer`. When
-        used with a non-linear search the `output_path` points to the search's results folder and this function
+        Images are output to the `image` folder of the `image_path` in a subfolder called `tracer`. When
+        used with a non-linear search the `image_path` points to the search's results folder and this function
         visualizes the maximum log likelihood `Tracer` inferred by the search so far.
 
         Visualization includes individual images of attributes of the tracer (e.g. its image, convergence, deflection
@@ -148,8 +148,8 @@ class PlotterInterface(AgPlotterInterface):
         Visualizes the positions of a model-fit, where these positions are used to resample lens models where
         the positions to do trace within an input threshold of one another in the source-plane.
 
-        Images are output to the `image` folder of the `output_path` in a subfolder called `positions`. When
-        used with a non-linear search the `output_path` points to the search's results folder.
+        Images are output to the `image` folder of the `image_path` in a subfolder called `positions`. When
+        used with a non-linear search the `image_path` points to the search's results folder.
 
         The visualization is an image of the strong lens with the positions overlaid.
