+++ autolens/quantity/model/analysis.py
@@ -1,19 +1,19 @@
-import os
 from typing import Optional
 
 import autofit as af
 import autogalaxy as ag
 
-from autogalaxy.quantity.model.plotter_interface import PlotterInterfaceQuantity
-
-from autolens.analysis.plotter_interface import PlotterInterface
+from autogalaxy.quantity.model.visualizer import VisualizerQuantity
 from autolens.analysis.analysis.lens import AnalysisLens
-from autogalaxy.quantity.plot.fit_quantity_plotters import FitQuantityPlotter
 from autolens.quantity.model.result import ResultQuantity
 from autolens.quantity.fit_quantity import FitQuantity
 
 
 class AnalysisQuantity(ag.AnalysisQuantity, AnalysisLens):
+
+    Result = ResultQuantity
+    Visualizer = VisualizerQuantity
+
     def __init__(
         self,
         dataset: ag.DatasetQuantity,
@@ -76,92 +76,3 @@ class AnalysisQuantity(ag.AnalysisQuantity, AnalysisLens):
         tracer = self.tracer_via_instance_from(instance=instance)
 
         return FitQuantity(dataset=self.dataset, tracer=tracer, func_str=self.func_str)
-
-    def visualize(
-        self,
-        paths: af.DirectoryPaths,
-        instance: af.ModelInstance,
-        during_analysis: bool,
-    ) -> None:
-        """
-        Output images of the maximum log likelihood model inferred by the model-fit. This function is called throughout
-        the non-linear search at regular intervals, and therefore provides on-the-fly visualization of how well the
-        model-fit is going.
-
-        The visualization performed by this function includes:
-
-        - Images of the best-fit `Plane`, including the images of each of its galaxies.
-
-        - Images of the best-fit `FitQuantity`, including the model-image, residuals and chi-squared of its fit to
-        the imaging data.
-
-        The images output by this function are customized using the file `config/visualize/plots.yaml`.
-
-        Parameters
-        ----------
-        paths
-            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
-            visualization, and the pickled objects used by the aggregator output by this function.
-        instance
-            An instance of the model that is being fitted to the data by this analysis (whose parameters have been set
-            via a non-linear search).
-        during_analysis
-            If True the visualization is being performed midway through the non-linear search before it is finished,
-            which may change which images are output.
-        """
-
-        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
-            return
-
-        fit = self.fit_quantity_for_instance(instance=instance)
-
-        plotter_interface = PlotterInterfaceQuantity(output_path=paths.image_path)
-        plotter_interface.fit_quantity(
-            fit=fit, fit_quanaity_plotter_cls=FitQuantityPlotter
-        )
-
-        plotter_interface = PlotterInterface(output_path=paths.image_path)
-        plotter_interface.tracer(
-            tracer=fit.tracer, grid=self.dataset.grid, during_analysis=during_analysis
-        )
-
-    def make_result(
-        self,
-        samples_summary: af.SamplesSummary,
-        paths: af.AbstractPaths,
-        samples: Optional[af.SamplesPDF] = None,
-        search_internal: Optional[object] = None,
-    ) -> ResultQuantity:
-        """
-        After the non-linear search is complete create its `ResultQuantity`, which includes:
-
-        - The samples of the non-linear search (E.g. MCMC chains, nested sampling samples) which are used to compute
-        the maximum likelihood model, posteriors and other properties.
-
-        - The model used to fit the data, which uses the samples to create specific instances of the model (e.g.
-        an instance of the maximum log likelihood model).
-
-        - The non-linear search used to perform the model fit.
-
-        The `ResultQuantity` object contains a number of methods which use the above objects to create the max
-        log likelihood `Tracer`, `FitQuantity`,etc.
-
-        Parameters
-        ----------
-        samples
-            The samples of the non-linear search, for example the MCMC chains.
-        search
-            The non-linear search used to perform this model-fit.
-
-        Returns
-        -------
-        ResultQuantity
-            The result of fitting the model to the imaging dataset, via a non-linear search.
-        """
-        return ResultQuantity(
-            samples_summary=samples_summary,
-            paths=paths,
-            samples=samples,
-            search_internal=search_internal,
-            analysis=self,
-        )
