+++ test_autolens/analysis/test_plotter_interface.py
@@ -20,7 +20,7 @@ def test__tracer(
     if os.path.exists(plot_path):
         shutil.rmtree(plot_path)
 
-    plotter_interface = vis.PlotterInterface(output_path=plot_path)
+    plotter_interface = vis.PlotterInterface(image_path=plot_path)
 
     plotter_interface.tracer(
         tracer=tracer_x2_plane_7x7, grid=masked_imaging_7x7.grid, during_analysis=False
@@ -50,7 +50,7 @@ def test__image_with_positions(
     if os.path.exists(plot_path):
         shutil.rmtree(plot_path)
 
-    plotter_interface = vis.PlotterInterface(output_path=plot_path)
+    plotter_interface = vis.PlotterInterface(image_path=plot_path)
 
     plotter_interface.image_with_positions(image=image_7x7, positions=positions_x2)
